Setup Command.java file to prepare for methods refactoring
Two default methods, getMandatoryTags() and getOptionalTags() is
declared for the commands that do not have mandatory tags or optional
tags. The commands that have either of each will override the method
to return a string array of their corresponding tags.

A number of setters are also created to allow polymorphism.