Omit boarding time for passenger input
Boarding time have been automated to be set 45 mins before the departure
time, this would make it bug free and prevent the manager from inputting
a baording time that is later than the departure time.

I have also the function of the delay to also update the boarding time
accordingly when departure time is updated. This would ensure that the
passenger and flight list are in sync.